multi: Move signal to internal package.
