+++ internal/signal/signal_syscall.go
@@ -5,7 +5,7 @@
 
 //go:build windows || aix || android || darwin || dragonfly || freebsd || hurd || illumos || ios || linux || netbsd || openbsd || solaris
 
-package main
+package signal
 
 import (
 	"syscall"
